Convert ETL into Python package
