+++ candigETL/mapping/mappings.py
